fix: moved checkFileExists , readFileContent, renameFileIfNecessary to file.service.ts
